# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ecommerce Playground API",
    "description": "API para sistema de e-commerce de farmacia (Arquitectura de Servicios Refactorizada v2.1)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sales/{cart_id}/complete": {
      "post": {
        "tags": [
          "Sales"
        ],
        "summary": "Complete Sale",
        "description": "Completar una venta desde un carrito.\nEsto actualiza el inventario y crea el registro de venta.",
        "operationId": "complete_sale_api_sales__cart_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cart Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentDetails"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sale"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/{cart_id}/cancel": {
      "post": {
        "tags": [
          "Sales"
        ],
        "summary": "Cancel Sale",
        "description": "Anular un carrito pendiente (no revierte stock).",
        "operationId": "cancel_sale_api_sales__cart_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cart Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Cancel Sale Api Sales  Cart Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/daily": {
      "get": {
        "tags": [
          "Sales"
        ],
        "summary": "Get Daily Sales",
        "description": "Obtener el reporte de ventas del d\u00eda actual.",
        "operationId": "get_daily_sales_api_sales_daily_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReport"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sales/close-day": {
      "post": {
        "tags": [
          "Sales"
        ],
        "summary": "Close Day",
        "description": "Cerrar el d\u00eda y generar el resumen final de ventas.\n(Cumple con el subm\u00f3dulo de cierre de 'single-e-commerce-demo.docx')",
        "operationId": "close_day_api_sales_close_day_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Close Day Api Sales Close Day Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/admin/users/": {
      "get": {
        "tags": [
          "User Admin"
        ],
        "summary": "Get All Users",
        "description": "Obtiene una lista de todos los usuarios en el sistema.\n(Solo para 'admin')",
        "operationId": "get_all_users_api_admin_users__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "title": "Response Get All Users Api Admin Users  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User Admin"
        ],
        "summary": "Create New User",
        "description": "Crea un nuevo usuario (empleado) en el sistema.\n(Solo para 'admin')",
        "operationId": "create_new_user_api_admin_users__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{user_id}": {
      "get": {
        "tags": [
          "User Admin"
        ],
        "summary": "Get User By Id",
        "description": "Obtiene los detalles de un usuario espec\u00edfico por ID.\n(Solo para 'admin')",
        "operationId": "get_user_by_id_api_admin_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User Admin"
        ],
        "summary": "Update User",
        "description": "Actualiza el perfil de un usuario (roles, estado, nombre, email).\n(Solo para 'admin')",
        "operationId": "update_user_api_admin_users__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/reset-password": {
      "post": {
        "tags": [
          "User Admin"
        ],
        "summary": "Admin Reset Password",
        "description": "Permite a un admin forzar una nueva contrase\u00f1a para cualquier usuario.\n(Solo para 'admin')",
        "operationId": "admin_reset_password_api_admin_users_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Admin Reset Password Api Admin Users Reset Password Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminPasswordResetRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "new_password": {
            "type": "string",
            "minLength": 8,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "new_password"
        ],
        "title": "AdminPasswordResetRequest",
        "description": "DTO para que un Admin fuerce una nueva contrase\u00f1a"
      },
...
      "CartItem": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quantity"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "product_name",
          "quantity",
          "price"
        ],
        "title": "CartItem",
        "description": "Item individual en un carrito de compras."
      },
...
      "DailyReport": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "sales_count": {
            "type": "integer",
            "title": "Sales Count"
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total"
          },
          "sales": {
            "items": {
              "$ref": "#/components/schemas/Sale"
            },
            "type": "array",
            "title": "Sales"
          }
        },
        "type": "object",
        "required": [
          "date",
          "sales_count",
          "total",
          "sales"
        ],
        "title": "DailyReport",
        "description": "DTO para el reporte de ventas diario."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentDetails": {
        "properties": {
          "payment_method": {
            "type": "string",
            "title": "Payment Method"
          },
          "payment_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Type"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "bank": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          }
        },
        "type": "object",
        "required": [
          "payment_method"
        ],
        "title": "PaymentDetails"
      },
...
      "Sale": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "cart_id": {
            "type": "string",
            "title": "Cart Id"
          },
          "customer_name": {
            "type": "string",
            "title": "Customer Name"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CartItem"
            },
            "type": "array",
            "title": "Items"
          },
          "currency_id": {
            "type": "string",
            "title": "Currency Id"
          },
          "payment_details": {
            "$ref": "#/components/schemas/PaymentDetails"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "completed"
          },
          "completed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed By"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          },
          "region_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Id"
          },
          "subtotal": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Subtotal"
          },
          "tax_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Tax Amount"
          },
          "igtf_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Igtf Amount",
            "default": 0.0
          },
          "total_with_tax": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total With Tax"
          }
        },
        "type": "object",
        "required": [
          "cart_id",
          "customer_name",
          "customer_id",
          "items",
          "currency_id",
          "payment_details",
          "subtotal",
          "tax_amount",
          "total_with_tax"
        ],
        "title": "Sale"
      },
...
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username"
          },
          "password_hash": {
            "type": "string",
            "title": "Password Hash"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted",
            "default": false
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password_hash"
        ],
        "title": "User",
        "description": "Usuario del sistema (Admin).\nREFACTOR: 'role' es ahora 'roles' (una lista) para RBAC."
      },
      "UserCreateRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username"
          },
          "plain_password": {
            "type": "string",
            "minLength": 8,
            "title": "Plain Password"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "description": "Ej: ['sales', 'products']"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "username",
          "plain_password"
        ],
        "title": "UserCreateRequest",
        "description": "DTO para que un Admin cree un nuevo usuario"
      },
...
      "UserUpdateRequest": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "roles": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roles"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "UserUpdateRequest",
        "description": "DTO para que un Admin O un usuario actualice datos (sin contrase\u00f1a)"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}